Create anshu
Stress has become an inseparable part of our daily existence and to know its symptoms is essential for everyone including parents and children.

The term “stress” encompasses a variety of situations, each affecting individuals differently. From the frustration of navigating through traffic to the pressures of exams or dealing with health concerns, stress manifests in diverse ways.

School children, teenagers and college students , everyone is facing a lot of stress these days. Therefore is is crucial or parents to guide their children how to deal with day to day stresses.

This article aims to help parents to explore the concept of stress, its symptoms, and shed light on the physiological responses our bodies undergo during challenging situations of their children.